proj/requirements.txt
amqp==2.5.2
gevent==1.4.0  
billiard==3.6.3.0
celery==4.4.2
certifi==2019.11.28
chardet==3.0.4
idna==2.9
importlib-metadata==1.5.0
kombu==4.6.8
lxml==4.5.0
pymongo==3.10.1
pytz==2019.3
requests==2.23.0
urllib3==1.25.8
vine==1.3.0
zipp==3.1.0
python-dateutil==2.8.0 
